AIHubBackend/requirements.txt
fastapi
uvicorn
pydantic
requests
azure-functions
azure-identity
